Better global pref api
